Ajustes na gravação dos pagamentos da ficha

+++ src/database/DBPagamento.java
@@ -5,7 +5,6 @@
  */
 package database;
 
-import java.sql.Timestamp;
 import java.util.Date;
 
 /**
@@ -20,7 +19,7 @@ public class DBPagamento {
     private String strDataReferencia = "";
     @DBMExcluido
     private Date dataReferencia = null;
-    private float valor = 0;
+    private double valor = 0;
     private String strDataPagamento = "";
     @DBMExcluido
     private Date dataPagamento = null;
@@ -54,11 +53,11 @@ public class DBPagamento {
         strDataReferencia = Conexao.getInstance().dateToString(dataReferencia);
     }
 
-    public float getValor() {
+    public double getValor() {
         return valor;
     }
 
-    public void setValor(float valor) {
+    public void setValor(double valor) {
         this.valor = valor;
     }
